pyproject.toml
[tool.poetry]
name = "est8"
version = "1.0.0"
description = "A card-based estate building game."
authors = ["MartinHowarth <howarth.martin@gmail.com>"]
license = "MIT"
readme = 'README.md'

repository = "https://github.com/MartinHowarth/est8"
homepage = "https://github.com/MartinHowarth/est8"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.9"
cocos2d = "^0.6.7"
shimmer = "^1.0.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
pydocstyle = "^5.0.1"
mypy = "^0.750"
pytest = "^5.3.1"
pytest-subtests = "^0.2.1"
mock = "^3.0.5"
pytest-mock = "^1.13.0"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

[tool.pytest]
mock_use_standalone_module = true